refactor: separate `errors::format` and other functions to `components/format`
