+++ src/error.cpp
@@ -18,17 +18,3 @@ Error make(const std::string& msg) {
 }
 
 }  // namespace error
-
-namespace fmt {
-
-format_parse_context::iterator formatter<errors::Error>::parse(
-    format_parse_context& ctx) const {
-  return ctx.begin();
-}
-
-format_context::iterator formatter<errors::Error>::format(
-    const errors::Error& err, format_context& ctx) const {
-  return format_to(ctx.out(), "error: {}", err.message());
-}
-
-}  // namespace fmt
